backend/requirements.txt
fastapi 
uvicorn 
qpython 
openai 
pydantic
langchain
langchain-openai
python-dotenv